More siddur calendaring and make siddur explanations into comments

+++ __tests__/api_request_handler_runner.ts
@@ -1,12 +1,22 @@
 import {exec} from "child_process";
+import {ArgumentParser} from "argparse";
 import {writeJson} from "../util/json_files";
 import {ApiRequestHandler} from "../api_request_handler";
 import {RecordingRequestMaker, testPages} from "./api_request_handler_base";
 
-exec("rm test_data/api_request_handler/*");
+const argParser = new ArgumentParser({description: "Update api_request_handler test files"});
+argParser.add_argument('--title');
+argParser.add_argument('--page');
+const flags = argParser.parse_args();
+
+if (!flags.title && !flags.page) {
+  exec("rm test_data/api_request_handler/*");
+}
 
 const requestHandler = new ApiRequestHandler(new RecordingRequestMaker());
 for (const testPage of testPages) {
+  if (flags.title && flags.title !== testPage.title) continue;
+  if (flags.page && flags.page !== testPage.page) continue;
   requestHandler.handleRequest(testPage.title, testPage.page)
     .then(results => writeJson(testPage.outputFilePath(), results))
     .catch(e => {
